// src/models/movie.ts
export default class Movie {
  adult: string = '';
  title: string = '';
  name: string = '';
  overview: string = '';
  genre_ids: Array<number> = Array<number>();
  first_air_date: string = '';
  genres: string = '';
  id: number = 0;
  vote_average: string = '';
  vote_count: number = 0;
  video: boolean = false;
  poster_path: string = '';
  plot: string = '';
  original_title: string = '';
  original_name: string = '';
  original_language: string = '';
  release_date: string = '';
  backdrop_path: string = '';
  popularity: string = '';

}